+++ src/bin/src/errors.rs
@@ -1,4 +1,3 @@
-use thiserror::Error;
 use ferrumc_core::errors::CoreError;
 use ferrumc_ecs::errors::ECSError;
 use ferrumc_events::errors::EventsError;
@@ -7,6 +6,7 @@ use ferrumc_plugins::errors::PluginsError;
 use ferrumc_storage::errors::StorageError;
 use ferrumc_utils::errors::UtilsError;
 use ferrumc_world::errors::WorldError;
+use thiserror::Error;
 
 #[derive(Debug, Error)]
 pub enum BinaryError {
@@ -19,7 +19,7 @@ pub enum BinaryError {
     #[error("Events error: {0}")]
     Events(#[from] EventsError),
 
-    #[error("Net error: {0}")] 
+    #[error("Net error: {0}")]
     Net(#[from] NetError),
 
     #[error("Plugins error: {0}")]
@@ -33,11 +33,10 @@ pub enum BinaryError {
 
     #[error("World error: {0}")]
     World(#[from] WorldError),
-    
+
     #[allow(dead_code)]
     #[error("{0}")]
     Custom(String),
-
     // #[error("IO error: {0}")]
     // Io(#[from] std::io::Error),
