+++ src/bin/src/lib.rs
