+++ src/lib/adapters/nbt/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 
 thiserror = { workspace = true }
 hashbrown = { workspace = true }
+libflate = { workspace = true }
 
 [lints]
 workspace = true
