+++ src/lib/adapters/nbt/src/errors.rs
