+++ src/lib/core/src/errors.rs
