+++ src/lib/derive_macros/src/events/mod.rs
@@ -1,24 +1,24 @@
-use proc_macro::{TokenStream};
+use proc_macro::TokenStream;
 use quote::{format_ident, quote};
-use syn::{parse_macro_input, Expr, Lit, Meta};
 use syn::punctuated::Punctuated;
 use syn::token::Comma;
+use syn::{parse_macro_input, Expr, Lit, Meta};
 
 pub fn event_handler_fn(attr: TokenStream, input: TokenStream) -> TokenStream {
     let args = parse_macro_input!(attr with Punctuated::<Meta, syn::Token![,]>::parse_terminated);
     let input = parse_macro_input!(input as syn::ItemFn);
 
     let fn_name = &input.sig.ident;
-    
+
     let priority = parse_priority(args);
 
     let register_fn_name = format_ident!("__register_listener_{}", fn_name);
-    
+
     let event_type = extract_event_type(&input).ty;
 
     let output = quote! {
         #input
-        
+
         #[ctor::ctor]
         fn #register_fn_name() {
             // ::ferrumc_events::infrastructure::insert_into_events(
@@ -28,7 +28,7 @@ pub fn event_handler_fn(attr: TokenStream, input: TokenStream) -> TokenStream {
             );
         }
     };
-    
+
     output.into()
 }
 
@@ -70,7 +70,6 @@ fn parse_priority(args: Punctuated<Meta, Comma>) -> u8 {
                     }
                     _ => panic!("Expected a string literal for the priority attribute. Possible values are: fastest, fast, normal, slow, slowest, or values between 0 and 255. Where 0 is the fastest and 255 is the slowest")
                 };
-
             }
             _ => {}
         }
@@ -81,17 +80,17 @@ fn parse_priority(args: Punctuated<Meta, Comma>) -> u8 {
 
 fn extract_event_type(input: &syn::ItemFn) -> syn::PatType {
     let inputs = &input.sig.inputs;
-    
+
     if inputs.len() != 1 {
         panic!("Expected the event handler to have exactly one argument (the event)");
     }
-    
+
     let syn::FnArg::Typed(pat_type) = &inputs[0] else {
         panic!("Expected the first argument to be a typed pattern");
     };
-    
+
     // let syn::Path { segments, .. } = &type_path.path;
-    
+
     pat_type.clone()
 }
