+++ src/lib/net/src/errors.rs
@@ -1,6 +1,6 @@
-use thiserror::Error;
 use ferrumc_net_encryption::errors::NetEncryptionError;
 use ferrumc_net_packets::errors::NetPacketError;
+use thiserror::Error;
 
 #[derive(Debug, Clone, Error)]
 pub enum NetError {
