+++ src/lib/plugins/src/errors.rs
