+++ src/lib/storage/src/errors.rs
