+++ src/lib/utils/logging/src/errors.rs
