+++ src/lib/utils/profiling/src/errors.rs
